requirements.txt

pandas
matplotlib
scikit-learn
google-cloud-bigquery
